Border Detection with Pixel Conversion
Makes the tick marks on a ruler and the lesion black and then find those
borders (and puts them into an array). The cell array with the largest
border is obviously the mole, so it returns that border. With the rest
of the borders, it finds the centroids of the first two tick marks and
uses the Pythagorean theorem to calculate the distance. This works in
case the ruler is crooked or something. It also assumes that all the
tick marks on a ruler are uniform.